+++ src/net/peacefulcraft/sco/items/CraftingValidSlotItem.java
@@ -1,13 +1,10 @@
-package net.peacefulcraft.sco.items.utilityitems;
+package net.peacefulcraft.sco.items;
 
 import java.util.ArrayList;
 
 import org.bukkit.Material;
 
-import net.peacefulcraft.sco.items.ItemIdentifier;
-import net.peacefulcraft.sco.items.ItemTier;
-
-public class CraftingInvalidSlot implements ItemIdentifier {
+public class CraftingValidSlotItem implements ItemIdentifier {
 
     @Override
     public String getName() { return ""; }
@@ -16,7 +13,7 @@ public class CraftingInvalidSlot implements ItemIdentifier {
     public ArrayList<String> getLore() { return new ArrayList<String>(); }
 
     @Override
-    public Material getMaterial() { return Material.RED_STAINED_GLASS_PANE; }
+    public Material getMaterial() { return Material.GREEN_STAINED_GLASS_PANE; }
 
     private Integer quantity;
         @Override
@@ -24,7 +21,6 @@ public class CraftingInvalidSlot implements ItemIdentifier {
 
         @Override
         public void setQuantity(Integer quantity) { this.quantity = quantity; }
-
 
     @Override
     public ItemTier[] getAllowedTiers() { return new ItemTier[]{ ItemTier.COMMON }; }
@@ -36,9 +32,9 @@ public class CraftingInvalidSlot implements ItemIdentifier {
     public boolean isDroppable() { return false; }
 
     @Override
-    public boolean isMovable() { return false; }
+    public boolean isMovable() { return true; }
 
-    public CraftingInvalidSlot(ItemTier tier, Integer quantity) {
+    public CraftingValidSlotItem(ItemTier tier, Integer quantity) {
         this.quantity = quantity;
     }
 }
